+++ internal/broker/broker.go
@@ -18,6 +18,7 @@ var ValidTopics = map[string]bool{
 	"user.query":         true,
 	"group.query":        true,
 	"ansible.progress":   true,
+	"service.query":      true,
 }
 
 // Broker is a thread-safe, topic-based pub/sub message broker.